vinyl: store iid, space_id and path to the index in the log
We can't remove run files right after index removal, because that would
break recovery from WAL. So we can only do cleanup on snapshot. By that
time, all associated structures have been freed, which makes it
impossible to find the index directory. Therefore we need to store the
information required to lookup the index directory in the metadata log.
This information will be used for garbage collection.

Needed for #1860