vinyl: log run file creation in the metadata journal
Currently, we write info about a run to the metadata log after it was
created. As a result, if run creation fails due to a write error and we
fail to delete the incomplete file for some reason, we won't be able
to look it up and retry deletion. To fix that, we need to log a record
containing info necessary for the run file deletion before starting to
write a run file. This patch adds a new type of log record for this
purpose, VY_LOG_PREPARE_RUN. It contains index and run IDs and is logged
when a dump/compact/split task is created. For now, on recovery we add
all incomplete runs, i.e. those runs that have VY_LOG_PREPARE_RUN but
don't have matching VY_LOG_INSERT_RUN to per index ->incomplete_runs
list, and ignore them on log rotation. We will use it later to retry run
file deletion.

Needed for #2029